Fixed major memory leaks in equation_system.
This required some changes. It turned-out that the mfem::Array used to store kernels in the hephaestus field maps were not destroying the kernels correctly so I have switched to using std::vectors containing unique pointers for each kernel.

Additionally, I have added typedefs since the names were becoming too long and added virtual destructors to the Kernel and derived classes to ensure that the correct destructor is called for subclasses.